// mobile/src/services/generated/api/models/PersonCreateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type PersonCreateDto = {
  id?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  contactNumber?: string | null;
  streetName?: string | null;
  emailAddress?: string | null;
  city?: string | null;
  postalCode?: string | null;
  country?: string | null;
  dateOfBirth?: string | null;
};