// mobile/src/services/generated/api/models/SkillDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { RefListSeniorityRating } from './RefListSeniorityRating';

export type SkillDto = {
  id?: string;
  name?: string | null;
  yearsOfExperience?: number | null;
  seniorityRating?: RefListSeniorityRating;
  seniorityRatingName?: string | null;
  personId?: string | null;
};